+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.3.1 - 2021-03-24
+### Added
+- Add `PointOfInterestTypeAppender` which allows for mods to add blocks to
+  vanilla POI types' block state list.
+
 ## v0.3.0 - 2021-03-11
 ### Added
 - Add `PlumeCraftingTableBlock` which allows for non-vanilla crafting table